bash completion moved to separate file
also added kubectl completion

+++ bin/setup/git-completion.sh
@@ -1,10 +1,11 @@
 #!/usr/bin/env bash
 SRC=https://raw.githubusercontent.com/git/git/master/contrib/completion/git-completion.bash
 DST=~/bin/.git-completion.bash
-mkdir -p ~/bin/
 # http://apple.stackexchange.com/questions/55875/have-git-autocomplete-branches-at-the-command-line
-curl $SRC -o $DST
-
+if [ ! -f $DST ]; then
+  mkdir -p ~/bin/
+  curl $SRC -o $DST
+fi
 #Then I added to my ~/.bash_profile file the following 'execute if it exists' code:
 
 #DST=~/bin/.git-completion.bash
